requirements.txt
python-telegram-bot==21.6
anthropic>=0.28.0
pdfplumber>=0.11.0
pypdf>=4.0.0
